Final project/requirements.txt
pytest
selenium
selenium-wire
webdriver-manager
allure-pytest
blinker==1.4